vimana_photoscan_automation_ customized/requirements.txt
piexif
Pillow
pandas
pymediainfo
nose
sphinx
networkx
matplotlib
pytz
pymysql
requests
urllib3
charted
idna